Fixed some errors, created new separate class based on sample code in teleop, will use it to determine which version of the TeamCode method I should keep.
